+++ packages/dex-core/src/exmergo_dex_core/explore/__init__.py
@@ -1,0 +1,9 @@
+"""Explore capability: read-only landscape understanding.
+
+Turns a connection into a ranked, profiled map of the warehouse: inventory and
+ranking to make drill-down selective, column profiling with PII flags, and
+declared/inferred relationships. ``commands`` holds the CLI orchestrators; the
+sibling modules are the engine they drive.
+"""
+
+from __future__ import annotations
